// ms-merchant/src/models/requests/search-merchant.request.ts
export class SearchMerchantBy {
    id: number;
    mId: number;
    companyName: string;
    transactionEmail: string;
    userId: number;
    contactName: string;
    contactPhone: string;
    contactEmail: string;
    paymentMethod: string;
    merchantStatus: string;
    status: number;
    createdBy: number;
    take: number;
    skip: number;
}